Multi-threaded server with multiple clients
Implemented the server to be thread safe with a thread for each client that connects to it. Implemented server part of the new networking model using a thread for each client. Implemented name control at lobby creation. Created PreGame and PreGameController to handle game creation with distributed MVC pattern, RemoteView in working.
Client need to be updated